pyproject.toml
[project]
name = "memo-chatbot"
version = "0.1.0"
description = "메모 관리 챗봇 with MCP"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2.0,<0.3.0",
    "langchain>=0.3.0,<0.4.0",
    "langchain-openai>=0.2.0,<0.3.0",
    "langchain-core>=0.3.37,<0.4.0",
    "mcp>=1.0.0",
    "langchain-mcp>=0.2.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["chatbot", "mcp-server"]
